Squashed commit of the following:
commit 6ca0d5d5c5f67670d17635eab4bc0d098944e74e
Author: Eric Chen <159481679+echen-adobe@users.noreply.github.com>
Date:   Mon Aug 11 17:24:15 2025 -0500

    Update simplified-pricing-cards-v2.css

commit 6c073cae5c9900576b5f26251447f1da6db0ed0a
Author: Eric Chen <159481679+echen-adobe@users.noreply.github.com>
Date:   Mon Aug 11 17:23:12 2025 -0500

    Update image-tooltip.js

commit 3a09402cea9aea4e4e2b42f69807151b1967636c
Author: Eric Chen <159481679+echen-adobe@users.noreply.github.com>
Date:   Mon Aug 11 17:22:02 2025 -0500

    Update image-tooltip.js

commit 2c5ed9444259be09b2aa5f163fda8a3f08c384fc
Author: Eric Chen <159481679+echen-adobe@users.noreply.github.com>
Date:   Mon Aug 11 17:10:32 2025 -0500

    ...
